refactor(gui-app): keep the typing dots private to AgentSpinningDots
`WorkingDots` was exported as a second spinner primitive and the chat run
indicator and the rate-limit "Refreshing" text imported it directly, around
the rule that `AgentSpinningDots` is the only spinner seam. The dots are now
a module-private implementation of its existing `typing` variant and both
call sites render `<AgentSpinningDots variant="typing">`, so cadence,
reduced-motion and pane-visibility behaviour cannot diverge between spinner
APIs.
